requirements.txt
discord2
python-dateutil
requests
beautifulsoup4